+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 3.77.0 (Unreleased)
+
+### Added
+- Support resource discovery for `streaming` resources
+
 ## 3.76.0 (May 19, 2020)
 
 ### Added